Add 'getRedis' method.
Redis::scan method takes the first arg by reference, so we have a 'Parameter 1 to Redis::scan() expected to be a reference, value given' error when run it through proxy __call here. Run Redis::scan directly is the easiest way to solve this issue therefore we need a getter method for redis parameter.